requirements.txt
ollama
pydantic
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
pyyaml
